backend/pyproject.toml
[tool.poetry]
name = "invoicecash_aggregator"
version = "0.1.0"
description = ""
authors = ["Serghei <serghei@oveit.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.1.2"
Flask-Security = "^3.0.0"
Flask-Cors = "^3.0.10"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
Flask-Mail = "^0.9.1"
Flask-JWT-Extended = "^4.4.1"
Flask-RQ2 = "^18.3"
Flask-RESTful = "^0.3.9"
flasgger = "^0.9.5"
flask-marshmallow = "^0.14.0"
marshmallow = "^3.16.0"
marshmallow-jsonapi = "^0.24.0"
marshmallow-sqlalchemy = "^0.28.0"
psycopg2-binary = "^2.9.3"
typing-extensions = "^4.2.0"
importlib-resources = "^5.7.1"
email-validator = "^1.2.1"
redis = "^4.3.3"
rq = "^1.10.1"
gunicorn = "^20.1.0"
python-dotenv = "^0.20.0"
apispec = "^5.2.2"
Werkzeug = "^2.1.2"
pycrypto = "^2.6.1"
pyOpenSSL = "^22.0.0"
Flask-Admin = "^1.6.0"

[tool.poetry.dev-dependencies]
unittest2 = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
